Убрать логику работы с хранилищем данных в соответсвующие классы
